backend/requirements.txt
vocode==0.1.111a1
uvicorn==0.21.1
python-dotenv
